MAINT: STY: Consolidate test_helper with test_fftpack.
The tests in test_helper.py are moved to test_fftpack.py and
test_helper.py deleted. The test_fftpack.py file is made PEP8 compliant.
